controle de fluxos

+++ api-rest/app/Http/Controllers/EsporteController.php
@@ -7,6 +7,9 @@ use Illuminate\Http\Request;
 
 class EsporteController extends Controller
 {
+    public function __construct(Esporte $esporte) {
+       $this->esporte = $esporte; 
+    }
     /**
      * Display a listing of the resource.
      *
@@ -14,7 +17,8 @@ class EsporteController extends Controller
      */
     public function index()
     {
-        $esportes = Esporte::all();
+        //$esportes = Esporte::all();
+        $esportes = $this->esporte->all();
         return $esportes;
     }
 
@@ -36,19 +40,24 @@ class EsporteController extends Controller
      */
     public function store(Request $request)
     {
-        $esporte = Esporte::create($request->all());
-        dd($esporte);
+        //$esporte = Esporte::create($request->all());
+        //dd($esporte);
+        $esporte = $this->esporte->create($request->all());
         return $esporte;
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  \App\Models\Esporte  $espaco
+     * @param  Integer
      * @return \Illuminate\Http\Response
      */
-    public function show(Esporte $esporte)
+    public function show($id)
     {
+        $esporte = $this->esporte->find($id);
+        if($esporte === null) {
+            return ['erro' => 'Recurso Indisponível'];
+        }
         return $esporte;
     }
 
@@ -67,16 +76,17 @@ class EsporteController extends Controller
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\Esporte  $espaco
+     * @param  Integer
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Esporte $esporte)
+    public function update(Request $request, $id)
     {
-        /*
-        print_r($request->all()); //dados atualizados
-        echo '<hr>';
-        print_r($esporte->getAttributes()); //dados antigos
-        */
+        $esporte = $this->esporte->find($id);
+
+        if($esporte === null) {
+            return ['erro' => 'Não foi possível realizar a atualização. Recurso indisponível'];
+        }
+
         $esporte->update($request->all());
         return $esporte;
     }
@@ -84,11 +94,17 @@ class EsporteController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Models\Esporte  $espaco
+     * @param  Integer
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Esporte $esporte)
+    public function destroy($id)
     {
+        $esporte = $this->esporte->find($id);
+
+        if($esporte === null) {
+            return ['erro' => 'Não foi possível realizar a atualização. Recurso indisponível'];
+        }
+
         $esporte->delete();
         return ['msg' => 'O esporte foi removido com sucesso!'];
     }
